requirements.txt
flask==2.3.3
google-generativeai==0.3.2
numpy==1.24.3
pandas==2.0.3
python-dotenv==1.0.0
gunicorn==21.2.0
openpyxl==3.1.2
httpx==0.24.1 